created the order route
